💥 reorganize; rename Setter/Getter toolkits to Reader/Writer
